Move TestAddressRepository to test fixtures
COAND-911
